Split up parquet::arrow::array_reader
